fix graph restart
