+++ sematic/graph.py
@@ -13,8 +13,8 @@ from sematic.db.models.factories import get_artifact_value
 from sematic.db.models.run import Run
 from sematic.resolvers.type_utils import make_list_type, make_tuple_type
 from sematic.storage import Storage
+from sematic.utils.algorithms import breadth_first_search, topological_sort
 from sematic.utils.memoized_property import memoized_indexed, memoized_property
-from sematic.utils.sorting import breadth_first, topological_sort
 
 RunID = str
 RunsByID = Dict[RunID, Run]
@@ -208,7 +208,8 @@ class Graph:
             ]
             for run_id in layer_run_ids
         }
-        return topological_sort(dependencies)
+        sorted = topological_sort(dependencies)
+        return sorted
 
     def _reverse_execution_order(self, layer_run_ids: List[RunID]) -> List[RunID]:
         """
@@ -259,7 +260,24 @@ class Graph:
             for parent_id, runs in self._runs_by_parent_id.items()
         }
 
-        return breadth_first(layers, run_sorter)
+        run_ids = []
+
+        start_nodes = [r.id for r in self._runs_by_parent_id[None]]
+
+        def get_next(run_id):
+            sorted = run_sorter(layers.get(run_id, []))
+            return sorted
+
+        def visit(run_id):
+            run_ids.append(run_id)
+
+        breadth_first_search(
+            start=start_nodes,
+            get_next=get_next,
+            visit=visit,
+        )
+
+        return run_ids
 
     @memoized_indexed
     def _get_run_ancestor_ids(self, run_id: RunID) -> List[RunID]:
