+++ sematic/utils/BUILD
@@ -42,12 +42,6 @@ sematic_py_lib(
 )
 
 sematic_py_lib(
-    name = "sorting",
-    srcs = ["sorting.py"],
-    deps = [":algorithms"],
-)
-
-sematic_py_lib(
     name = "json",
     srcs = ["json.py"],
     deps = [],
