+++ sematic/utils/algorithms.py
@@ -1,6 +1,6 @@
 # Standard Library
 from collections import deque
-from typing import Callable, Hashable, List, Optional, TypeVar
+from typing import Callable, Dict, Hashable, List, Optional, TypeVar
 
 T = TypeVar("T")
 
@@ -37,3 +37,42 @@ def breadth_first_search(
         visit(visiting)
         visited_keys.add(key)
         to_visit.extend(get_next(visiting))
+
+
+def topological_sort(
+    dependencies: Dict[str, List[Optional[str]]],
+) -> List[str]:
+    """
+    Sorts a list of strings topologically.
+
+    Parameters
+    ----------
+    dependencies: Dict[str, List[Optional[str]]]
+        A mapping of node to its dependencies. The root nodes have [None] as dependencies.
+
+    Returns
+    -------
+    List[str]
+        IDs sorted topologically.
+    """
+    sorted_ids: List[str] = []
+
+    def remove_nones(li):
+        return [element for element in li if element is not None]
+
+    def visit(id_):
+        sorted_ids.append(id_)
+
+    def find_next_ids(previous_id):
+        # Sorting for determinism.
+        next_ids = sorted(
+            element for element, deps in dependencies.items() if previous_id in deps
+        )
+        return next_ids
+
+    start_nodes = [
+        element for element, deps in dependencies.items() if remove_nones(deps) == []
+    ]
+
+    breadth_first_search(start=start_nodes, get_next=find_next_ids, visit=visit)
+    return sorted_ids
