--- sematic/utils/tests/test_sorting.py
@@ -1,14 +1,0 @@
-# Sematic
-from sematic.utils.sorting import topological_sort
-
-
-def test_topological_sort():
-    dependencies = {
-        "A": [None],
-        "B": ["A"],
-        "C": ["A"],
-        "D": ["C", "E"],
-        "E": ["B", "A"],
-    }
-
-    assert topological_sort(dependencies) == ["A", "B", "C", "E", "D"]
